Move libc_library_list back into repo.
