+++ snapcraft/plugins/copy.py
@@ -39,8 +39,6 @@ import os
 import glob
 
 import snapcraft
-from snapcraft.internal import libraries
-
 
 logger = logging.getLogger(__name__)
 
@@ -117,7 +115,7 @@ def _link_or_copy(source, destination, boundary):
         normalized = os.path.normpath(os.path.join(destination_dirname, link))
         if os.path.isabs(link) or not normalized.startswith(boundary):
             # Only follow symlinks that are NOT pointing at libc (LP: #1658774)
-            if link not in libraries.libc_library_list():
+            if link not in snapcraft.repo.libc_library_list():
                 follow_symlinks = True
 
     snapcraft.common.link_or_copy(source, destination,
