+++ snapcraft/tests/plugins/test_copy.py
@@ -26,7 +26,6 @@ from snapcraft.plugins.copy import (
     CopyPlugin,
     _recursively_link
 )
-from snapcraft.internal import libraries
 from snapcraft import tests
 
 
@@ -336,7 +335,7 @@ class TestCopyPlugin(tests.TestCase):
 
         # Even though this symlink is absolute, since it's to libc the copy
         # plugin shouldn't try to follow it or modify it.
-        libc_libs = libraries.libc_library_list()
+        libc_libs = snapcraft.repo.libc_library_list()
 
         # We don't care which lib we're testing with, as long as it's a .so.
         libc_library_path = [lib for lib in libc_libs if '.so' in lib][0]
